move proxy.js to lib/Proxy.js
